requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg[binary]==3.3.1
Pillow==10.4.0

# Storage & Media
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# Phone & OTP
django-phonenumber-field==8.0.0
phonenumbers==8.13.29
requests==2.32.3

# Security & Utils
django-ratelimit==4.1.0
django-axes==6.5.0
django-otp==1.5.0
qrcode==7.4.2
python-dotenv==1.0.1

# CORS
django-cors-headers==4.4.0

# Audit Log
django-auditlog==3.0.0
